introduced proper package structure
