+++ microquant/__init__.py
@@ -1,5 +1,7 @@
 """Top-level package for MicroQuant."""
 
-__author__ = """Johannes Müller"""
-__email__ = 'johannes.mueller88@hotmail.de'
-__version__ = '0.1.0'
+__version__ = '0.0.1'
+
+from . import segmentation
+from . import registration
+from . import measurement
