+++ setup.cfg
@@ -1,18 +1,36 @@
-[bumpversion]
-current_version = 0.1.0
-commit = True
-tag = True
+[metadata]
+name = MicroQuant
+version = 0.0.1
+author = Johannes Müller
+author_email = johannes_richard.mueller@tu-dresden.de
 
-[bumpversion:file:setup.py]
-search = version='{current_version}'
-replace = version='{new_version}'
+license = BSD-3-Clause
+description = Collection of analysis function for pairs of H&E and immunofluorescent image data
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/jo-mueller/MicroQuant.git
+classifiers =
+    Development Status :: 2 - Pre-Alpha
+    Intended Audience :: Developers
+    Topic :: Software Development :: Testing
+    Programming Language :: Python
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
+    Operating System :: OS Independent
+    License :: OSI Approved :: BSD License
 
-[bumpversion:file:microquant/__init__.py]
-search = __version__ = '{current_version}'
-replace = __version__ = '{new_version}'
 
-[bdist_wheel]
-universal = 1
+[options]
+packages = find:
+python_requires = >=3.7
 
-[flake8]
-exclude = docs
+# add your package requirements here
+install_requires =
+    numpy
+    torch
+    scikit-image
+    scikit-learn
+    tqdm
+    tifffile
