+++ setup.py
@@ -1,44 +1,5 @@
 #!/usr/bin/env python
+# -*- coding: utf-8 -*-
+from setuptools import setup
 
-"""The setup script."""
-
-from setuptools import setup, find_packages
-
-with open('README.rst') as readme_file:
-    readme = readme_file.read()
-
-with open('HISTORY.rst') as history_file:
-    history = history_file.read()
-
-requirements = [ ]
-
-test_requirements = [ ]
-
-setup(
-    author="Johannes Müller",
-    author_email='johannes.mueller88@hotmail.de',
-    python_requires='>=3.6',
-    classifiers=[
-        'Development Status :: 2 - Pre-Alpha',
-        'Intended Audience :: Developers',
-        'License :: OSI Approved :: BSD License',
-        'Natural Language :: English',
-        'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.6',
-        'Programming Language :: Python :: 3.7',
-        'Programming Language :: Python :: 3.8',
-    ],
-    description="Does some stuff",
-    install_requires=requirements,
-    license="BSD license",
-    long_description=readme + '\n\n' + history,
-    include_package_data=True,
-    keywords='microquant',
-    name='microquant',
-    packages=find_packages(include=['microquant', 'microquant.*']),
-    test_suite='tests',
-    tests_require=test_requirements,
-    url='https://github.com/jo-mueller/microquant',
-    version='0.1.0',
-    zip_safe=False,
-)
+setup()
